+++ python/components/deframers/ccsds_concatenated_deframer.py
@@ -23,15 +23,17 @@ class ccsds_concatenated_deframer(gr.hier_block2, options_block):
 
     Args:
         frame_size: frame size (not including parity check bytes) (int)
-        differential: whether to use differential coding (bool)
-        dual_basis: use dual basis instead of conventional (bool)
-        use_scrambler: use CCSDS synchronous scrambler (bool)
-        nasa_dsn: use NASA-DSN instead of CCSDS/NASA-GSFC convolutional code (bool)
+        precoding: either None or 'differential' for differential precoding (str)
+        rs_basis: Reed-Solomon basis, either 'conventional' or 'dual' (str)
+        rs_interleaving: Reed-Solomon interleaving depth (int)
+        scrambler: scrambler to use, either 'CCSDS' or 'none' (str)
+        convolutional: convolutional code to use, one of the following:
+            'CCSDS', 'NASA-DSN', 'CCSDS uninverted', 'NASA-DSN uninverted' (str)
         syncword_threshold: number of bit errors allowed in syncword (int)
         options: Options from argparse
     """
-    def __init__(self, frame_size = 223, differential = False, dual_basis = False,
-                     use_scrambler = True, nasa_dsn = False,
+    def __init__(self, frame_size = 223, precoding = None, rs_basis = 'dual',
+                     rs_interleaving = 1, scrambler = 'CCSDS', convolutional = 'CCSDS',
                      syncword_threshold = None, options = None):
         gr.hier_block2.__init__(self, "ccsds_concatenated_deframer",
             gr.io_signature(1, 1, gr.sizeof_float),
@@ -41,16 +43,16 @@ class ccsds_concatenated_deframer(gr.hier_block2, options_block):
         self.message_port_register_hier_out('out')
 
         self.delay1 =  blocks.delay(gr.sizeof_float, 1)
-        self.viterbi0 = ccsds_viterbi(nasa_dsn = nasa_dsn)
-        self.viterbi1 = ccsds_viterbi(nasa_dsn = nasa_dsn)
+        self.viterbi0 = ccsds_viterbi(convolutional)
+        self.viterbi1 = ccsds_viterbi(convolutional)
         self.char2float0 = blocks.char_to_float(1, 1)
         self.char2float1 = blocks.char_to_float(1, 1)
         self.addconst0 = blocks.add_const_ff(-0.5)
         self.addconst1 = blocks.add_const_ff(-0.5)
-        self.rs0 = ccsds_rs_deframer(frame_size, differential, dual_basis,
-                                         use_scrambler, syncword_threshold, options)
-        self.rs1 = ccsds_rs_deframer(frame_size, differential, dual_basis,
-                                         use_scrambler, syncword_threshold, options)
+        self.rs0 = ccsds_rs_deframer(frame_size, precoding, rs_basis, rs_interleaving,
+                                         scrambler, syncword_threshold, options)
+        self.rs1 = ccsds_rs_deframer(frame_size, precoding, rs_basis, rs_interleaving,
+                                         scrambler, syncword_threshold, options)
 
         self.connect(self, self.viterbi0, self.char2float0, self.addconst0, self.rs0)
         self.connect(self, self.delay1, self.viterbi1, self.char2float1, self.addconst1, self.rs1)